Relocate test/utils/solidity.py to nucypher/utilities/ethereum.py
