+++ tests/contracts/base/test_multisig.py
@@ -15,13 +15,12 @@ You should have received a copy of the GNU Affero General Public License
 along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-
 import pytest
 from eth_tester.exceptions import TransactionFailed
 from eth_utils import to_canonical_address
 
 from nucypher.blockchain.eth.constants import NULL_ADDRESS
-from tests.utils.solidity import to_32byte_hex
+from nucypher.utilities.ethereum import to_32byte_hex
 
 
 def sign_hash(testerchain, account: str, data_hash: bytes) -> dict:
